resolution du bug liée à la gestion de la dette

+++ frontend/lib/model/flux_financier/tranche_payement_credit.dart
@@ -9,7 +9,7 @@ class TranchePayementModel {
 
   factory TranchePayementModel.fromJson(Map<String, dynamic> json) {
     return TranchePayementModel(
-       montantPaye: json["montantPaye"],
+      montantPaye: json["montantPaye"].toDouble(),
       datePayement: DateTime.fromMillisecondsSinceEpoch(json['datePayement']),
     );
   }
